Refactor lexer and update related data structures
The lexer code was moved to parser/lexer.c from lexer/lexer.c and refactored, improving readability and efficiency in command line word processing tasks. The data structure s_words was renamed to s_word and updated in minishell.h to better serve the lexer function. An addition to the data structure t_ast was also made for parsing operations, improving the programmability and functionality of the mini shell.